# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 浯島財務管理系統 — Integration API
  description: |
    通用對外整合 API。讓 PM、POS、ERP 等外部系統可推送收入 / 支出資料進入本系統。

    詳細串接說明：see [`docs/integration-api.md`](integration-api.md)
  version: "2.0.0"
  contact:
    name: 大哉實業
    url: https://money.homi.cc
servers:
  - url: https://money.homi.cc
    description: 生產環境
  - url: http://localhost:5001
    description: 本地開發

# ─────────────────────────────────────────────
# 三類安全方案
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer Token 驗證。authType=token 或 both 時必須提供。
        Token 由管理員在後台建立來源時設定。
    HmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC-SHA256 簽章。authType=hmac 或 both 時必須提供。
        格式：`sha256=<hex>`，用 webhookSecret 對 raw body 簽。
  schemas:
...
    IntegrationEvent:
      type: object
      properties:
        id:
          type: integer
        integrationType:
          type: string
          enum: [income, expense]
        sourceId:
          type: integer
        sourceKey:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        httpMethod:
          type: string
        httpPath:
          type: string
        statusCode:
          type: integer
        outcome:
          type: string
          enum: [success, auth_failed, validation_failed, duplicate, error, retried]
        errorMessage:
          type: string
          nullable: true
        latencyMs:
          type: integer
          nullable: true
        attempt:
          type: integer
        parentEventId:
          type: integer
          nullable: true
        linkedWebhookId:
          type: integer
          nullable: true
        requestIp:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/integrations/events:
    get:
      summary: 查詢拋接紀錄
      tags: [Integration]
      parameters:
        - name: integrationType
          in: query
          schema: { type: string, enum: [income, expense] }
        - name: sourceId
          in: query
          schema: { type: integer }
        - name: sourceKey
          in: query
          schema: { type: string }
        - name: direction
          in: query
          schema: { type: string, enum: [inbound, outbound] }
        - name: outcome
          in: query
          schema:
            type: string
            enum: [success, auth_failed, validation_failed, duplicate, error, retried]
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: until
          in: query
          schema: { type: string, format: date-time }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: 分頁結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/IntegrationEvent"
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
  /api/integrations/events/{id}:
    get:
      summary: 取得單筆事件詳情
      tags: [Integration]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 事件詳情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationEvent"
        "404":
          description: 事件不存在
  /api/integrations/events/{id}/replay:
    post:
      summary: Replay — 將指定事件的 payload 重新送一次
      description: 用原本的 source 設定重新計算簽章並送出；結果寫入新 event，parent_event_id 指向原 event。
      tags: [Integration]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: 已重送
          content:
            application/json:
              schema:
                type: object
                properties:
                  replayed: { type: boolean }
                  originalEventId: { type: integer }
                  newEvent:
                    $ref: "#/components/schemas/IntegrationEvent"
                  result:
                    type: object
...
  /api/integrations/sources/{type}/{id}/test:
    post:
      summary: 產生測試 payload（可選實際送出）
      description: |
        後端依 source 的 fieldMapping 產生 sample payload + 正確簽章。
        executeForReal=true 時會真的走完整 webhook 流程。
      tags: [Integration]
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string, enum: [income, expense] }
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                executeForReal:
                  type: boolean
                  default: false
      responses:
        "200":
          description: 測試結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  sourceKey: { type: string }
                  payload: { type: object }
                  headers: { type: object }
                  curl: { type: string, description: 可複製貼上的 curl 指令 }
                  executed: { type: boolean }
                  result:
                    type: object
                    nullable: true
                    description: executeForReal=true 時提供